Split freeorion_tools to separate files
