+++ default/python/AI/freeorion_tools/__init__.py
@@ -1,13 +1,11 @@
 from ._freeorion_tools import (
-    assertion_fails, cache_by_turn_persistent, cache_for_current_turn,
-    cache_for_session, chat_human, dict_to_tuple, get_ai_tag_grade, get_partial_visibility_turn, get_ship_part,
+    assertion_fails, chat_human, dict_to_tuple, get_ai_tag_grade,
+    get_partial_visibility_turn, get_ship_part,
     get_species_tag_grade,
     ReadOnlyDict, tech_is_complete,
-    tuple_to_dict, UserString, UserStringList, ppstring,
+    tuple_to_dict, ppstring,
 )
-from .extend_freeorion_AI_interface import patch_interface
-from .timers import AITimer
-from freeorion_tools.fo_chat_handler import configure_debug_chat, process_chat_message
+
 from freeorion_tools.combine_ratings import combine_ratings
 from freeorion_tools._profile import profile
 from freeorion_tools._fleet_position import get_fleet_position
