+++ default/python/AI/turn_state/_dry_dock_state.py
@@ -3,7 +3,8 @@ from typing import Mapping, Tuple
 import AIDependencies
 import freeOrionAIInterface as fo
 
-from freeorion_tools import cache_for_current_turn, ReadOnlyDict
+from freeorion_tools import ReadOnlyDict
+from freeorion_tools.caching import cache_for_current_turn
 
 
 @cache_for_current_turn
