+++ default/python/AI/turn_state/_empire_resources.py
@@ -1,7 +1,7 @@
 import AIDependencies
 import freeOrionAIInterface as fo
 from EnumsAI import FocusType
-from freeorion_tools import cache_for_current_turn
+from freeorion_tools.caching import cache_for_current_turn
 from turn_state._planet_state import _get_planets_info
