+++ default/python/AI/turn_state/_planet_state.py
@@ -2,7 +2,8 @@ from typing import Callable, FrozenSet, List, Mapping, Set, Tuple
 
 import AIDependencies
 import freeOrionAIInterface as fo
-from freeorion_tools import cache_for_current_turn, ReadOnlyDict
+from freeorion_tools import ReadOnlyDict
+from freeorion_tools.caching import cache_for_current_turn
 
 
 class PlanetInfo:
