+++ default/python/AI/turn_state/_supply_state.py
@@ -2,7 +2,7 @@ from logging import error, warning
 from typing import Dict, Mapping, Tuple
 
 import freeOrionAIInterface as fo
-from freeorion_tools import cache_for_current_turn
+from freeorion_tools.caching import cache_for_current_turn
 
 
 @cache_for_current_turn
